requirements.txt
torch>=2.0
pyyaml>=6.0
tqdm>=4.66
spacy>=3.7
matplotlib>=3.7
sacrebleu>=2.4

# 如果要用spacy_model分词,需要使用spacy的语言模型:
# python -m spacy download en_core_web_sm
# python -m spacy download de_core_news_sm